+++ small_text/integrations/pytorch/datasets.py
@@ -1,5 +1,6 @@
 import numpy as np
 
+from abc import ABC
 from small_text.base import LABEL_UNLABELED
 from small_text.data import Dataset, DatasetView
 from small_text.data.datasets import check_size
@@ -13,10 +14,7 @@ except ModuleNotFoundError:
     raise PytorchNotFoundError('Could not import torchtext')
 
 
-class PytorchDataset(Dataset):
-
-    def __init__(self, device=None):
-        self.device = device
+class PytorchDataset(ABC):
 
     def to(self, other, non_blocking=False, copy=False):
         raise NotImplementedError()
@@ -87,7 +85,7 @@ class PytorchTextClassificationDataset(PytorchDataset):
     INDEX_TEXT = 0
     INDEX_LABEL = 1
 
-    def __init__(self, data, vocab, multi_label=False, target_labels=None, device=None):
+    def __init__(self, data, vocab, multi_label=False, target_labels=None):
         """
         Parameters
         ----------
@@ -105,8 +103,6 @@ class PytorchTextClassificationDataset(PytorchDataset):
             e.g. due to dataset splits, where the labels should however be considered by
             entities such as the classifier. If `None`, the target labels will be inferred
             from the labels encountered in `self.data`.
-        device : str or torch.device
-            Torch device on which the computation will be performed.
         """
         self._data = data
         self._vocab = vocab
@@ -119,13 +115,6 @@ class PytorchTextClassificationDataset(PytorchDataset):
         else:
             self.track_target_labels = True
             self._infer_target_labels()
-
-        if device is None:
-            self.device = None if len(data) == 0 else data[0][self.INDEX_TEXT].device
-        else:
-            self.device = device
-
-        super().__init__(device=device)
 
     def _infer_target_labels(self):
         inferred_target_labels = self._get_flattened_unique_labels()
@@ -252,8 +241,7 @@ class PytorchTextClassificationDataset(PytorchDataset):
             target_labels = None if self.track_target_labels else self._target_labels
             # TODO: clone vocab
             vocab = self._vocab
-            return PytorchTextClassificationDataset(data, vocab, target_labels=target_labels,
-                                                    device=self.device)
+            return PytorchTextClassificationDataset(data, vocab, target_labels=target_labels)
         else:
             self._data = data
             return self